Display a basic navbar and styling on the page
Website will need blogpost pages so I crated one to be used as a template. It was impossible to navigate between the homepage and the new blogpost page so i created a basic navbar. Style looked like crappy default HTML so I added basic CSS for readability and design points. The design mockup PDF was bad because I keep changing it so I replaced it with a Figma link in the README.